Use file mtime to determine if a file is ready to be backed up.
We previously relied on implementation details of the HDFS filesystem to
determine if file was ready for backing up by looking at the
construction status of a file's blocks. This meant that any alternative
HDFS implementations, such as MapR's, were not compatible. Using the
mtime gives a very similar user experience, since a user can choose
whether or not a file should be included in a backup based on how much
mtime drift they configure. Due to this behavioral change, the minor
version has been increased to 0.10.