Updating to Hadoop from CDH 4.4.0
Modified HdfsTest to use the new MiniDFSCluster builder pattern.

There is a terrible, awful hack in setupHdfsClusters. For some unknown reason,
the first write to HDFS fails under this version of MiniDFSCluster. All
subsequent writes work fine. It does not create the proper folders on the
filesystem on the first write, but they are created on the second write.